[32/64-bit] - Getting 32/64-bit builds working
As part of the push to get 32/64 bit firing up more completely, HELLO as a project can now be built
using either:

make 32-bit, or
make 64-bit

The .list output files suggest things make general sense. This will be tested against a complete 32-bit
deployment to ensure the 32-bit versions of everything are working as expected. Once that is in place, the
64-bit versions of everything can start to be integrated with the 64-bit kernel.